Se a traspasado los métodos para Admin que estaban en Usuario y comentar nuevos posibles métodos
